Initial attempt to consolidate packages into core.
