+++ packages/core/src/constants.ts
