+++ packages/core/src/services/index.ts
@@ -1,3 +1,4 @@
-export * from './provider/ProviderService';
-export * from './persistent-storage.service';
-export * from './logging.service';
+export * from "./provider/provider.service";
+export * from "./persistent-storage/persistent-storage.service";
+export * from "./logger/logging.service";
+export * from "./event-emitter/event-emitter.service";
