+++ packages/core/src/wallet/index.ts
@@ -1,0 +1,2 @@
+export * from "./wallet";
+export * from "./actions";
