--- packages/utils/README.md
@@ -1,7 +1,0 @@
-# utils
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test utils` to execute the unit tests via [Jest](https://jestjs.io).
