--- packages/utils/jest.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  displayName: "utils",
-  preset: "../../jest.preset.js",
-  globals: {
-    "ts-jest": {
-      tsconfig: "<rootDir>/tsconfig.spec.json",
-    },
-  },
-  transform: {
-    "^.+\\.[tj]sx?$": "ts-jest",
-  },
-  moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
-  coverageDirectory: "../../coverage/packages/utils",
-};
