--- packages/utils/project.json
@@ -1,34 +1,0 @@
-{
-  "root": "packages/utils",
-  "sourceRoot": "packages/utils/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nrwl/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/packages/utils",
-        "main": "packages/utils/src/index.ts",
-        "tsConfig": "packages/utils/tsconfig.lib.json",
-        "buildableProjectDepsInPackageJsonType": "dependencies",
-        "assets": ["packages/utils/*.md"]
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["packages/utils/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["coverage/packages/utils"],
-      "options": {
-        "jestConfig": "packages/utils/jest.config.js",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": ["utils"]
-}
