+++ workspace.json
@@ -2,7 +2,6 @@
   "version": 2,
   "projects": {
     "core": "packages/core",
-    "utils": "packages/utils",
     "ledger-wallet": "packages/ledger-wallet",
     "math-wallet": "packages/math-wallet",
     "near-wallet": "packages/near-wallet",
